Using "notifyOf..." naming convention in ApplicationEvents.
HandlingEvents now have a unique sequence number, and CargoUpdater looks events based on that to update the cargo in question.

First version of pessimistic locking of cargo during rerouting.

Cargo is now a facade for aggregate-internal classes Delivery and Projections.

Delivery calculates current voyage, last known location and transport status on the fly.

Introduced a tracking id generator domain service for getting next tracking id, instead of doing it in the cargo repository.